move vm in-operator test from spec to unit_node
