--- tests/specs/node/vm_in_operator_walks_prototype/__test__.jsonc
@@ -1,4 +1,0 @@
-{
-  "args": "run -A main.cjs",
-  "output": "main.out"
-}
