--- tests/specs/node/vm_in_operator_walks_prototype/main.out
@@ -1,3 +1,0 @@
-true
-true
-false
